refactor: split ir/compile.ts into focused modules, iterative topoSort, temp file cleanup
- Split compile.ts (386 lines) into: constants.ts, helpers.ts, validate.ts,
  discover.ts, and a thin compile.ts orchestrator (61 lines)
- Split execute.ts: extract runStep dispatch into run-step.ts, remove dead
  transcribe/silence-detect ops from the switch
- Rewrite topo-sort.ts from recursive DFS to iterative Kahn algorithm
  (prevents stack overflow on large plans)
- Add temp file cleanup in extractAudio/detectSilence via try/finally
  (preserves URL inputs, deletes toTempFile outputs on success and failure)
- Validate nested providerOptions.varg.provider.generate_audio in compile
  (prevents audio native from being silently dropped on gateway models)
- Await rejects.toThrow in execute test to fix async matcher
